instance_segmentator/requirements.txt
streamlit==1.32.0
ultralytics==8.1.24
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless==4.9.0.80
numpy>=1.24.0
pillow>=10.2.0
pyyaml==6.0.1
openai==1.14.2
python-dotenv==1.0.1
